pyproject.toml
[project]
name = "discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord.py",
    "pynacl>=1.5.0",
    "python-dotenv",
    "yt-dlp>=2025.8.11",
]
